refactor: move cli update modules into cli/update/ subpackage
Move _update, _update_checks, _update_checks_fetch, _update_checks_source
to cli/update/ and update all internal and external importers.
