+++ src/autoskillit/cli/update/__init__.py
@@ -1,0 +1,4 @@
+"""CLI update and version-check utilities."""
+
+from autoskillit.cli.update._update import run_update_command  # noqa: F401
+from autoskillit.cli.update._update_checks import run_update_checks  # noqa: F401
